+++ base64Hash/base64Hash_test.go
@@ -1,12 +1,12 @@
-package shortHash
+package base64Hash
 
 import (
 	"testing"
 )
 
-func TestGenerateShortHash(t *testing.T) {
+func TestGenerateBase64Hash(t *testing.T) {
 	t.Run("Should return error when `length` is not multiple of 4", func(t *testing.T) {
-		hash, err := GenerateShortHash(3)
+		hash, err := GenerateBase64Hash(3)
 
 		if hash == "" && err == nil {
 			t.Errorf("Function didn't return error for incorrect input")
@@ -14,13 +14,13 @@ func TestGenerateShortHash(t *testing.T) {
 	})
 
 	t.Run("Should generate unique hash every time function is called", func(t *testing.T) {
-		hash1, err := GenerateShortHash(4)
+		hash1, err := GenerateBase64Hash(4)
 
 		if len(hash1) != 4 && err != nil {
 			t.Errorf("Hash length should be 4 but is %v", len(hash1))
 		}
 
-		hash2, err := GenerateShortHash(8)
+		hash2, err := GenerateBase64Hash(8)
 
 		if len(hash2) != 8 && err != nil {
 			t.Errorf("Hash length should be 3 but is %v", len(hash2))
